Cloud/ML_model/requirements.txt
torch
torchvision
pillow
flask
flask-restful
requests
kafka-python
#opencv-python
#json
#jsonpickle